pyproject.toml
[project]
name = "boilermaker-servicebus"
version = "1.0.0a2"
description = "An async python Background task system using Azure Service Bus Queues"
authors = [{ "name" = "Erik Aker", "email" = "eaker@mulliganfunding.com" }]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aio-azure-clients-toolbox>=1.2.0",
    "rich >= 13.0.0",
    "anyio >= 4.13.0",
    "azure-core-tracing-opentelemetry >= 1.0.0b12",
    "azure-servicebus >= 7.14.3",
    "azure-identity >= 1.25.3",
    "opentelemetry-api >= 1.34.0",
    "pydantic >= 2.13.0",
    "pydantic-settings >= 2.13.1",
    "uuid-utils>=0.14.1",
]

[project.scripts]
boilermaker = "boilermaker.cli:main"

[project.urls]
Documentation = "https://mulliganfunding.github.io/boilermaker-servicebus/"
Repository = "https://github.com/MulliganFunding/boilermaker-servicebus"
Issues = "https://github.com/MulliganFunding/boilermaker-servicebus/issues"

[dependency-groups]
dev = [
    "build>=1.4.3",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.10",
    # Documentation dependencies
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.7.6",
    "mkdocstrings[python]>=1.0.3",
    "mkdocs-mermaid2-plugin>=1.2.3",
    "ty>=0.0.30",
]

[project.optional-dependencies]
repl = ["ipython"]

[build-system]
requires = ["setuptools", "versioningit"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
minversion = "8.3"
pythonpath = "."
asyncio_mode = "auto"
xfail_strict = true
addopts = "--import-mode=importlib --cov-report=term-missing:skip-covered --cov=boilermaker --strict-config --strict-markers --cov-report html"
testpaths = ["tests"]
filterwarnings = ["error"]

[tool.coverage.report]
exclude_also = [
    "def __repr__",
    "if self.debug",
    "raise AssertionError",
    "raise NotImplementedError",
    "@abstractmethod",
]

[tool.coverage.run]
omit = [
    # omit tracing.py and run
    "*/tracing.py",
]

[tool.ruff]
src = ["boilermaker", "tests"]
fix = true
show-fixes = true
output-format = "full"
line-length = 120

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # pyflakes
    "FA", # flake8-future-annotations
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
    "W",  # pycodestyle warning
]
ignore = ["N818", "N805"] # exception naming

[tool.ruff.lint.isort]
force-single-line = false
order-by-type = false
